requirements.txt
aiohttp~=3.11.11
python-dotenv==0.20.0
requests~=2.32.0
telegram~=0.0.1
djangorestframework~=3.15.2
django~=5.1.6
python-telegram-bot==20.1
httpx>=0.23.0
django-environ>=0.8.0
celery>=5.0
APScheduler~=3.11.0